# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hub Comercial iSolutis — API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/equipe": {
      "get": {
        "tags": [
          "Equipe"
        ],
        "summary": "Equipe",
        "description": "Pessoas da equipe (qualquer usuário logado): usado em responsáveis e na autoria dos registros.",
        "operationId": "equipe_api_v1_equipe_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MembroEquipe"
                  },
                  "type": "array",
                  "title": "Response Equipe Api V1 Equipe Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/usuarios": {
      "get": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Listar Usuarios",
        "operationId": "listar_usuarios_api_v1_usuarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioLeitura"
                  },
                  "type": "array",
                  "title": "Response Listar Usuarios Api V1 Usuarios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Criar Usuario",
        "operationId": "criar_usuario_api_v1_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCriar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/usuarios/{id_}": {
      "put": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Atualizar Usuario",
        "operationId": "atualizar_usuario_api_v1_usuarios__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Equipe",
          "Administração"
        ],
        "summary": "Remover Usuario",
        "operationId": "remover_usuario_api_v1_usuarios__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/faturamento": {
      "get": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_faturamento_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 2000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ano"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LancamentoLeitura"
                  },
                  "title": "Response Listar Api V1 Faturamento Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_faturamento_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LancamentoEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LancamentoLeitura"
                  },
                  "title": "Response Criar Api V1 Faturamento Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/faturamento/resumo": {
      "get": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Resumo",
        "operationId": "resumo_api_v1_faturamento_resumo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 2000,
              "title": "Ano"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoFaturamento"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/faturamento/exportar": {
      "get": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Exportar",
        "operationId": "exportar_api_v1_faturamento_exportar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 2000,
              "title": "Ano"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/faturamento/lote": {
      "post": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Criar Em Lote",
        "operationId": "criar_em_lote_api_v1_faturamento_lote_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaturamentoEmLote"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LancamentoLeitura"
                  },
                  "type": "array",
                  "title": "Response Criar Em Lote Api V1 Faturamento Lote Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/faturamento/{id_}": {
      "put": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_faturamento__id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LancamentoAtualizar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Excluir",
        "operationId": "excluir_api_v1_faturamento__id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/faturamento/{id_}/receber": {
      "post": {
        "tags": [
          "Faturamento"
        ],
        "summary": "Receber",
        "operationId": "receber_api_v1_faturamento__id___receber_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id "
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LancamentoLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FaturamentoEmLote": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "titulo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Titulo"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "projeto": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanoProjeto"
              },
              {
                "type": "null"
              }
            ]
          },
          "mensal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanoMensal"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cliente_id"
        ],
        "title": "FaturamentoEmLote",
        "description": "Geração atômica dos lançamentos de uma venda fechada (projeto parcelado e/ou mensalidades)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LancamentoEntrada": {
        "properties": {
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "projeto",
              "mensal",
              "consultoria",
              "outro"
            ],
            "title": "Tipo"
          },
          "descricao": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Descricao"
          },
          "valor": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,15}0*$)\\d{0,12}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Valor"
          },
          "vencimento": {
            "type": "string",
            "format": "date",
            "title": "Vencimento"
          },
          "status": {
            "type": "string",
            "enum": [
              "previsto",
              "recebido"
            ],
            "title": "Status",
            "default": "previsto"
          },
          "nf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nf"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "repetir": {
            "type": "integer",
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Repetir",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cliente_id",
          "tipo",
          "descricao",
          "valor",
          "vencimento"
        ],
        "title": "LancamentoEntrada"
      },
      "LancamentoLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          },
          "cliente_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cliente Id"
          },
          "cliente_nome": {
            "type": "string",
            "title": "Cliente Nome"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "descricao": {
            "type": "string",
            "title": "Descricao"
          },
          "valor": {
            "type": "number",
            "title": "Valor"
          },
          "vencimento": {
            "type": "string",
            "format": "date",
            "title": "Vencimento"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recebido_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recebido Em"
          },
          "nf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nf"
          },
          "negocio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negocio Id"
          },
          "orcamento_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orcamento Id"
          },
          "grupo_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grupo Id"
          },
          "parcela": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parcela"
          },
          "total_parcelas": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Parcelas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "criado_em",
          "atualizado_em",
          "criado_por",
          "atualizado_por",
          "versao",
          "cliente_id",
          "cliente_nome",
          "tipo",
          "descricao",
          "valor",
          "vencimento",
          "status",
          "recebido_em",
          "nf",
          "negocio_id",
          "orcamento_id",
          "grupo_id",
          "parcela",
          "total_parcelas"
        ],
        "title": "LancamentoLeitura"
      },
...
      "MembroEquipe": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email",
          "ativo"
        ],
        "title": "MembroEquipe",
        "description": "Visão mínima da equipe, disponível a qualquer pessoa logada (responsáveis, autoria)."
      },
...
      "ResumoFaturamento": {
        "properties": {
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "anos_disponiveis": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Anos Disponiveis"
          },
          "meses": {
            "items": {
              "$ref": "#/components/schemas/MesFaturamento"
            },
            "type": "array",
            "title": "Meses"
          },
          "total_recebido": {
            "type": "number",
            "title": "Total Recebido"
          },
          "total_previsto": {
            "type": "number",
            "title": "Total Previsto"
          },
          "total_recorrente": {
            "type": "number",
            "title": "Total Recorrente"
          }
        },
        "type": "object",
        "required": [
          "ano",
          "anos_disponiveis",
          "meses",
          "total_recebido",
          "total_previsto",
          "total_recorrente"
        ],
        "title": "ResumoFaturamento"
      },
...
      "UsuarioAtualizar": {
        "properties": {
          "versao": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Versao"
          },
          "nome": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nome"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo",
            "default": true
          },
          "senha": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Senha"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "versao",
          "nome",
          "admin"
        ],
        "title": "UsuarioAtualizar"
      },
      "UsuarioCriar": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Senha"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "title": "UsuarioCriar"
      },
      "UsuarioLeitura": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "senha_definida": {
            "type": "boolean",
            "title": "Senha Definida"
          },
          "ultimo_acesso": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Acesso"
          },
          "versao": {
            "type": "integer",
            "title": "Versao"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "nome",
          "admin",
          "ativo",
          "senha_definida",
          "ultimo_acesso",
          "versao"
        ],
        "title": "UsuarioLeitura"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}